requirements.txt
pandas>=1.3.0
sqlalchemy>=1.4.0
mysql-connector-python>=8.0.0
python-dotenv>=1.0.0